zss-app/backend/school_service/app/api/decorators.py
from functools import wraps
from flask import request, jsonify, current_app
import jwt

def token_required(f):
    @wraps(f)
    def decorated(*args, **kwargs):
        token = None
        if 'Authorization' in request.headers:
            token = request.headers['Authorization'].split(" ")[1]
        
        if not token:
            return jsonify({'message': 'Token nedostaje!'}), 401

        try:
            data = jwt.decode(token, current_app.config['SECRET_KEY'], algorithms=["HS256"])
            current_user = data
        except jwt.ExpiredSignatureError:
            return jsonify({'message': 'Token je istekao!'}), 401
        except jwt.InvalidTokenError:
            return jsonify({'message': 'Token je neispravan!'}), 401

        return f(current_user, *args, **kwargs)
    return decorated