zss-app/backend/school_service/requirements.txt
Flask==2.2.2
Flask-PyMongo==2.3.0
python-dotenv==0.21.0
PyJWT==2.6.0
Flask-Cors==3.0.10
requests==2.28.1
Werkzeug==2.3.8